requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
supabase==2.7.0
pydantic==2.9.0
python-multipart==0.0.9
httpx==0.27.0
google-generativeai==0.8.0
beautifulsoup4==4.12.0
dnspython==2.6.0
pandas==2.2.0
scikit-learn==1.5.0
xgboost==2.1.0
shap==0.45.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
